ci: rename and move footer component to landing page directory
